Change benchmark plugin layout so pytest loads it properly when running all tests (not necessary performance ones)
resolves #837
