+++ test_runner/conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = ("fixtures.zenith_fixtures")
+pytest_plugins = ("fixtures.zenith_fixtures", "fixtures.benchmark_fixture")
